+++ src/FormBuilder.js
@@ -20,6 +20,7 @@ import InnerCell from './Inputs'
 import {FaTrash as Trash} from 'react-icons/fa'
 import useTheme, {ThemeProvider} from './theme/useTheme'
 import {useAutoHideTooltip} from './useAutoHideTooltip'
+import ReactTooltip from 'react-tooltip'
 
 let inputEventListenerDebouncer = null
 
@@ -265,6 +266,7 @@ const FormBuilder = (props) => {
               expandable={expandable}
               setItemDimensions={setItemDimensions}
               dimensions={dimensions}
+              tooltipId={id}
             />
           </div>
         )
@@ -368,6 +370,10 @@ const FormBuilder = (props) => {
 
   dropItemDimensions.i = '-1'
 
+  useEffect(() => {
+    ReactTooltip.rebuild()
+  })
+
   return (
     <div
       id={id}
@@ -375,6 +381,13 @@ const FormBuilder = (props) => {
       ref={setContainerRef}
       style={style}
     >
+      <ReactTooltip
+        id={id}
+        getContent={content => {
+          // this is needed to maintain support for passing markup as a potential tooltip as data-tip in react-tooltip v4 - JRA 04/29/2026
+          return <div dangerouslySetInnerHTML={{__html: content}} />
+        }}
+      />
       <RGL
         ref={ReactGridLayout}
         autoSize={rglAutoSize}